Make Config standalone.
Almost everything config wise goes through Config
now. Also you can Load or Save it, and mutate the config.
Almost ready to remove syscfg entirely and start creating
the 5.4 style config.json, leaving config.dat to be
a write only legacy datafile.

+++ bbs/chat.cpp
@@ -1003,7 +1003,7 @@ int userinst(char* user) {
 }
 
 bool usercomp(const char* st1, const char* st2) {
-  for (int i = 0; i < size(st1); i++) {
+  for (size_t i = 0; i < size(st1); i++) {
     if (st1[i] != st2[i]) {
       return false;
     }
